LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/audio_loop_base-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

FAIL Check if audio.loop is set to true that expecting the seeking event is fired more than once assert_unreached: Reached unreachable code
